requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyee==9.0.4
pytz==2022.1
slackclient==2.9.4
slackeventsapi==3.0.1
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
